requirements.txt
aiogram==3.26.0
flask
requests
psycopg2-binary
